Fix excerpt error
A previous modification to this page accidentally disabled the 
functionality of using an excerpt in posts. The error was that the 
excerpt tag had no impact on how the post displayed on the homepage. 
Only summary and entire post were an option. This fixes that by checking 
if there's a summary, then checking if there's an error, and then if 
not, just posting the entire post to the homepage.