Add ability to display short posts
A new type of post, called a "short", is now allowed. The post does not
contain a title. As such, several changes were required which allowed
the post to be correctly displayed in index.html and on posts.html. The
title on the post page was also modified. In addition to these changes,
this commit adds a new template to facilitate these sorts of posts.