Allow image display in shorts
Previously, images could not be displayed in shorts. This was
because the img tag wasn't being properly removed. Changed
order of truncatewords and strip_html fixed this problem.
Updated comments to reflect that images can now be used with
shorts. Converted comments to liquid comments to prevent their
inclusion in output HTML.