Add sharing on social media
This commit adds the capability of sharing on social media. This is
accomplished by creating front-matter for adding text to be shared
and then corresponding feeds for each social media site. In addition,
this commit updates an example post by adding sharing front-matter
for testing purposes. This is based on the guide available at
https://www.jflh.ca/2017-02-08-how-i-automate-social-media-
sharing-of-my-jekyll-blog-articles.