Integrate updates from Blog
On posts page, allow shorts to pull the first 7 words instead of 5. Every
page has an automatically generated excerpt, so the final option
of just displaying the post content was never being reached. Removed
the note about the excerpt (since that's basically what page.content
truncated to 50 words does). If title is empty string, then if page.title
still evaluates to true. This places an empty string for the Facebook
and Twitter title. Changed to be if page.title != "", which will evaluate
to false if there is no title. This will then show the site title (Alex
Johnson).